pyproject.toml
[project]
name = "portfolioproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aif360>=0.6.1",
    "fairlearn>=0.12.0",
    "fpdf>=1.7.2",
    "imblearn>=0.0",
    "jupyter>=1.1.1",
    "kagglehub>=0.3.9",
    "langchain>=0.3.19",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.7",
    "matplotlib>=3.10.0",
    "numpy>=2.2.3",
    "openai>=1.64.0",
    "optuna>=4.2.1",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "streamlit>=1.42.2",
    "tabulate>=0.9.0",
]
